requirements.txt
streamlit==1.33.0
ultralytics==8.0.228
torch==2.1.2
torchvision==0.16.2
opencv-python==4.8.0.76
numpy==1.26.2
Pillow==10.1.0